[systemd/zsh] Use GPG-agent for SSH
Instead of separate SSH keys I'm migrating to using RSA subkeys of my
GPG key for SSH.

This modifies the gpg-agent unit to enable SSH and adds the
appropriate socket in .zshrc (todo: fish).

It is necessary to create a file "~/.gnupg/sshcontrol" with the keygrip
of the subkey used for SSH auth. The keygrip of keys can be seen with
`gpg --list-secret-keys --with-keygrip`